+++ forui/CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add `FThemeData.toApproximateMaterialTheme()`.
 * Add `FTimePickerStyle`.
 * Add `FPickerStyle.selectionHeightAdjustment`.
+* Add `FTileGroup.physics`.
 
 ### Changes